comment_service/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database dependencies
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication & Security (compatible with user_service)
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Text sanitization (XSS protection)
bleach==6.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0

# Development tools
python-dotenv==1.0.0
